KL-132/chore: move files
- 디렉토리 구조에 맞게 User 관련 파일 옮기기
- global의 response 디렉토리 common 하위로 옮기기
