+++ src/main/java/taco/klkl/domain/user/domain/User.java
@@ -11,7 +11,6 @@ import jakarta.persistence.PrePersist;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import taco.klkl.global.common.constants.UserConstants;
-import taco.klkl.global.common.enums.Gender;
 
 @Getter
 @NoArgsConstructor
