+++ src/test/java/taco/klkl/domain/user/dto/response/UserSimpleResponseDtoTest.java
@@ -5,8 +5,8 @@ import static org.assertj.core.api.Assertions.*;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
+import taco.klkl.domain.user.domain.Gender;
 import taco.klkl.domain.user.domain.User;
-import taco.klkl.global.common.enums.Gender;
 
 class UserSimpleResponseDtoTest {
 	@Test
